+++ command/cliconfig/cliconfig_test.go
@@ -1,4 +1,4 @@
-package main
+package cliconfig
 
 import (
 	"os"
