+++ command/cliconfig/testdata/credentials
