+++ command/cliconfig/testdata/hosts
